Create concrete UiConsumer classes for merge and export
